refactor(tools): remove obsolete Cursor shims
Why:
- Grok models now consume Pi native tool schemas and no longer require Cursor-compatible aliases.

What:
- Remove compatibility tool registration, scoping, delegation, and related dependencies.
- Keep image_gen preference synchronization across model and session changes.
- Remove obsolete shim documentation and tests.

Validation:
- bun run check (263 tests)
