+++ knip.json
@@ -1,5 +1,5 @@
 {
   "entry": ["src/shared/backup_worker.ts"],
   "project": ["**/*.ts", "!**/*.d.ts", "!.release-tools/**"],
-  "ignoreDependencies": ["ignore", "lint-staged", "minimatch"]
+  "ignoreDependencies": ["lint-staged"]
 }
